+++ packages/client/components/SnackbarMessageAction.tsx
@@ -1,24 +1,26 @@
-import React from 'react'
-import {SnackAction} from './Snackbar'
 import styled from '@emotion/styled'
-import PlainButton from './PlainButton/PlainButton'
-import {PALETTE} from '../styles/paletteV3'
+import React from 'react'
 import {DECELERATE} from '../styles/animation'
+import {PALETTE} from '../styles/paletteV3'
+import {Radius} from '../types/constEnums'
+import PlainButton from './PlainButton/PlainButton'
+import {SnackAction} from './Snackbar'
 
 interface Props {
   action: SnackAction | null | undefined
 }
 
 const Action = styled(PlainButton)({
-  borderRadius: 2,
+  borderRadius: Radius.SNACKBAR,
   color: PALETTE.ROSE_500,
   fontSize: 14,
   fontWeight: 600,
   marginLeft: 8,
   padding: 8,
+  backgroundColor: '#ffffff17',
   transition: `background 100ms ${DECELERATE}`,
   ':hover,:focus,:active': {
-    backgroundColor: '#ffffff0e'
+    backgroundColor: '#ffffff26'
   }
 })