+++ packages/server/database/types/DiscussStage.ts
@@ -17,5 +17,6 @@ export default class DiscussStage extends GenericMeetingStage {
     this.sortOrder = sortOrder
     this.discussionId = discussionId ?? generateUID()
     this.reflectionGroupId = reflectionGroupId
+    this.isNavigableByFacilitator = true
   }
 }