+++ packages/server/utils/analytics/analytics.ts
@@ -2,11 +2,12 @@ import {NewMeetingPhaseTypeEnum} from '../../database/types/GenericMeetingPhase'
 import Meeting from '../../database/types/Meeting'
 import MeetingMember from '../../database/types/MeetingMember'
 import MeetingTemplate from '../../database/types/MeetingTemplate'
+import {ReactableEnum} from '../../database/types/Reactable'
 import {IntegrationProviderServiceEnumType} from '../../graphql/types/IntegrationProviderServiceEnum'
 import {TeamPromptResponse} from '../../postgres/queries/getTeamPromptResponsesByIds'
-import {AnyMeeting} from '../../postgres/types/Meeting'
+import {AnyMeeting, MeetingTypeEnum} from '../../postgres/types/Meeting'
 import segment from '../segmentIo'
-import {createMeetingTemplateAnalyticsParams} from './helpers'
+import {createMeetingProperties} from './helpers'
 import {SegmentAnalytics} from './segment/SegmentAnalytics'
 
 export type OrgTierChangeEventProperties = {
@@ -23,6 +24,9 @@ export type AnalyticsEvent =
   | 'Meeting Started'
   | 'Meeting Joined'
   | 'Meeting Completed'
+  | 'Comment Added'
+  | 'Response Added'
+  | 'Reactji Interacted'
   // team
   | 'Integration Added'
   | 'Integration Removed'
@@ -57,7 +61,7 @@ class Analytics {
     )
     meetingMembers.forEach((meetingMember) => {
       const plaintextResponseContent = userIdsResponses[meetingMember.userId]
-      this.meetingEnd(meetingMember.userId, completedMeeting, meetingMembers, {
+      this.meetingEnd(meetingMember.userId, completedMeeting, meetingMembers, undefined, {
         responseAdded: !!plaintextResponseContent
       })
     })
@@ -75,12 +79,7 @@ class Analytics {
     template: MeetingTemplate
   ) => {
     meetingMembers.forEach((meetingMember) =>
-      this.meetingEnd(
-        meetingMember.userId,
-        completedMeeting,
-        meetingMembers,
-        createMeetingTemplateAnalyticsParams(template)
-      )
+      this.meetingEnd(meetingMember.userId, completedMeeting, meetingMembers, template)
     )
   }
 
@@ -90,12 +89,7 @@ class Analytics {
     template: MeetingTemplate
   ) => {
     meetingMembers.forEach((meetingMember) =>
-      this.meetingEnd(
-        meetingMember.userId,
-        completedMeeting,
-        meetingMembers,
-        createMeetingTemplateAnalyticsParams(template)
-      )
+      this.meetingEnd(meetingMember.userId, completedMeeting, meetingMembers, template)
     )
   }
 
@@ -103,22 +97,62 @@ class Analytics {
     userId: string,
     completedMeeting: Meeting,
     meetingMembers: MeetingMember[],
+    template?: MeetingTemplate,
     meetingSpecificProperties?: any
   ) => {
-    const {facilitatorUserId, meetingNumber, meetingType, phases, teamId} = completedMeeting
-    const presentMemberUserIds = meetingMembers.map(({userId}) => userId)
-    const wasFacilitator = userId === facilitatorUserId
     this.track(userId, 'Meeting Completed', {
-      hasIcebreaker: phases[0]?.phaseType === 'checkin',
-      // include wasFacilitator as a flag to handle 1 per meeting
-      wasFacilitator,
-      userIds: wasFacilitator ? presentMemberUserIds : undefined,
-      meetingType,
-      meetingNumber,
-      teamMembersCount: meetingMembers.length, // TODO: use team members count
-      teamMembersPresentCount: meetingMembers.length,
-      teamId,
+      wasFacilitator: completedMeeting.facilitatorUserId === userId,
+      ...createMeetingProperties(completedMeeting, meetingMembers, template),
       ...meetingSpecificProperties
+    })
+  }
+
+  meetingStarted = (userId: string, meeting: Meeting, template?: MeetingTemplate) => {
+    this.track(userId, 'Meeting Started', createMeetingProperties(meeting, undefined, template))
+  }
+
+  meetingJoined = (userId: string, meeting: Meeting) => {
+    this.track(userId, 'Meeting Joined', createMeetingProperties(meeting))
+  }
+
+  commentAdded = (userId: string, meeting: Meeting, isAnonymous, isAsync, isReply) => {
+    this.track(userId, 'Comment Added', {
+      meetingId: meeting.id,
+      meetingType: meeting.meetingType,
+      teamId: meeting.teamId,
+      isAnonymous,
+      isAsync,
+      isReply
+    })
+  }
+
+  responseAdded = (
+    userId: string,
+    meetingId: string,
+    teamPromptResponseId: string,
+    isUpdate: boolean
+  ) => {
+    this.track(userId, 'Response Added', {
+      meetingId,
+      teamPromptResponseId,
+      isUpdate
+    })
+  }
+
+  reactjiInteracted = (
+    userId: string,
+    meetingId: string,
+    meetingType: MeetingTypeEnum,
+    reactableId: string,
+    reactableType: ReactableEnum,
+    isRemove: boolean
+  ) => {
+    this.track(userId, 'Reactji Interacted', {
+      meetingId,
+      meetingType,
+      reactableId,
+      reactableType,
+      isRemove
     })
   }